Добавлен StatisticServiceProtocol и его реализация; подключён сервис статистики в контроллере
