+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -2,7 +2,7 @@ import UIKit
 
 final class MovieQuizViewController: UIViewController {
 
-    // MARK: - Outlets
+    // MARK: – View Outlets
 
     @IBOutlet private var imageView: UIImageView!
     @IBOutlet private var textLabel: UILabel!
@@ -10,22 +10,25 @@ final class MovieQuizViewController: UIViewController {
     @IBOutlet private var yesButton: UIButton!
     @IBOutlet private var noButton: UIButton!
 
-    // MARK: - Свойства
+    // MARK: – Свойства
 
     private lazy var alertPresenter = AlertPresenter(viewController: self)
 
+    private let statisticService: StatisticServiceProtocol = StatisticServiceImplementation()
+
     private var currentQuestionNumber = 0
+
     private var correctAnswers = 0
 
     private var questionFactory: QuestionFactoryProtocol!
+
     private var currentQuestion: QuizQuestion?
 
-    // MARK: - Lifecycle
+    // MARK: – Lifecycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Настройка UI
         imageView.layer.cornerRadius = 15
         imageView.layer.masksToBounds = true
 
@@ -39,15 +42,11 @@ final class MovieQuizViewController: UIViewController {
         questionFactory = factory
 
         showCurrentQuestion()
-        
-        // 4) Выводим путь к песочнице в консоль
-         print("📂 App Sandbox Path: \(NSHomeDirectory())")
-        
-        UserDefaults.standard.set(true, forKey: "viewDidLoad") 
 
+        print("📂 App Sandbox Path: \(NSHomeDirectory())")
     }
 
-    // MARK: - Actions
+    // MARK: – Actions
 
     @IBAction private func yesButtonClicked(_ sender: UIButton) {
         handleAnswer(true)
@@ -57,6 +56,7 @@ final class MovieQuizViewController: UIViewController {
         handleAnswer(false)
     }
 
+    // MARK: – Private methods
 
     private func showCurrentQuestion() {
         currentQuestionNumber += 1
@@ -69,9 +69,7 @@ final class MovieQuizViewController: UIViewController {
 
         guard let question = currentQuestion else { return }
         let isCorrect = (givenAnswer == question.correctAnswer)
-        if isCorrect {
-            correctAnswers += 1
-        }
+        if isCorrect { correctAnswers += 1 }
 
         showAnswerResult(isCorrect: isCorrect)
     }
@@ -99,7 +97,6 @@ final class MovieQuizViewController: UIViewController {
     }
 
     private func showAnswerResult(isCorrect: Bool) {
-        imageView.layer.masksToBounds = true
         imageView.layer.borderWidth = 8
         imageView.layer.borderColor = isCorrect
             ? UIColor.ypGreenIos.cgColor
@@ -117,7 +114,6 @@ final class MovieQuizViewController: UIViewController {
             buttonText: viewModel.buttonText
         ) { [weak self] in
             guard let self = self else { return }
-            // Сброс рамки
             self.imageView.layer.borderWidth = 0
             self.imageView.layer.borderColor = UIColor.clear.cgColor
 
@@ -155,24 +151,30 @@ final class MovieQuizViewController: UIViewController {
 }
 
 
+// MARK: – QuestionFactoryDelegate
 
 extension MovieQuizViewController: QuestionFactoryDelegate {
+
     func didReceiveNextQuestion(question: QuizQuestion?) {
         guard let question = question else {
             let total = questionFactory.totalQuestionsCount
+
+            statisticService.store(correct: correctAnswers, total: total)
+
             let dateFormatter = DateFormatter()
             dateFormatter.dateStyle = .short
             dateFormatter.timeStyle = .short
 
-            statisticService.store(correct: correctAnswers, total: total)
             let best = statisticService.bestGame
             let bestDateStr = dateFormatter.string(from: best.date)
+
             let message = """
             Ваш результат: \(correctAnswers)/\(total)
             Количество сыгранных квизов: \(statisticService.gamesCount)
             Рекорд: \(best.correct)/\(best.total) (\(bestDateStr))
             Средняя точность: \(String(format: "%.2f", statisticService.averageAccuracy))%
             """
+
             let resultViewModel = QuizResultsViewModel(
                 title: "Этот раунд окончен!",
                 text: message,
@@ -191,73 +193,4 @@ extension MovieQuizViewController: QuestionFactoryDelegate {
         }
     }
 }
-
-
-
-struct GameRecord: Codable {
-    let correct: Int
-    let total: Int
-    let date: Date
-}
-
-protocol StatisticService {
-    func store(correct count: Int, total questions: Int)
-    var gamesCount: Int { get }
-    var bestGame: GameRecord { get }
-    var averageAccuracy: Double { get }
-}
-
-final class StatisticServiceImplementation: StatisticService {
-    private let userDefaults = UserDefaults.standard
-
-    private enum Keys {
-        static let gamesCount      = "stat_gamesCount"
-        static let bestGame        = "stat_bestGame"
-        static let totalAccuracy   = "stat_totalAccuracy"
-    }
-
-    func store(correct count: Int, total questions: Int) {
-        let newGamesCount = gamesCount + 1
-        userDefaults.set(newGamesCount, forKey: Keys.gamesCount)
-
-        let thisAccuracy = Double(count) / Double(questions) * 100
-        let newTotalAcc = totalAccuracySum + thisAccuracy
-        userDefaults.set(newTotalAcc, forKey: Keys.totalAccuracy)
-
-        let currentBest = bestGame
-        if count > currentBest.correct {
-            let newRecord = GameRecord(
-                correct: count,
-                total: questions,
-                date: Date()
-            )
-            if let data = try? JSONEncoder().encode(newRecord) {
-                userDefaults.set(data, forKey: Keys.bestGame)
-            }
-        }
-    }
-
-    var gamesCount: Int {
-        userDefaults.integer(forKey: Keys.gamesCount)
-    }
-
-    private var totalAccuracySum: Double {
-        userDefaults.double(forKey: Keys.totalAccuracy)
-    }
-
-    var averageAccuracy: Double {
-        guard gamesCount > 0 else { return 0 }
-        return totalAccuracySum / Double(gamesCount)
-    }
-
-    var bestGame: GameRecord {
-        if let data = userDefaults.data(forKey: Keys.bestGame),
-           let record = try? JSONDecoder().decode(GameRecord.self, from: data) {
-            return record
-        }
-        return GameRecord(correct: 0, total: 0, date: Date())
-    }
-}
-
-private let statisticService: StatisticService = StatisticServiceImplementation()
