+++ MovieQuiz/Services/StatisticServiceProtocol.swift
@@ -1,0 +1,15 @@
+// StatisticServiceProtocol.swift
+// MovieQuiz
+
+import Foundation
+
+protocol StatisticServiceProtocol {
+    func store(correct count: Int, total questions: Int)
+
+    var gamesCount: Int { get }
+
+    var bestGame: GameRecord { get }
+
+    var averageAccuracy: Double { get }
+}
+
